Add register model decorator to eliminate eval in get_model
@register_model adds an entry to the model.model_builders dict, mapping the
name of the decorated function to the function itself. I replaced the eval in
get_models with a dictionary lookup. Additionally, registering each model
builder function makes it easy to require the --model-name flag only accepts
the name of valid models, so that the error message when passed an invalid
string is clearer.